Chart Staff, Perbaikan Ganti Password,eye icon di password field

+++ app/Http/Controllers/HolidayController.php
@@ -7,7 +7,7 @@ use Illuminate\Http\Request;
 use App\MasterHoliday;
 use Illuminate\Support\Facades\Auth;
 use RealRashid\SweetAlert\Facades\Alert;
-
+use Illuminate\Support\Facades\Gate;
 class HolidayController extends Controller
 {
     /**
@@ -17,6 +17,10 @@ class HolidayController extends Controller
      */
     public function index()
     {
+        if(Gate::denies('is_admin')){
+            return abort(403,'Access Denied, Only Admin Can Access');
+        }
+        elseif(Gate::allows('is_admin')) {
         $data = MasterHoliday::All();
         $user = Auth::user();
 
@@ -26,7 +30,7 @@ class HolidayController extends Controller
             'profile_photo'=>$user->profile_photo,
             'email'=>$user->email,
             'id'=>$user->id
-        ]);
+        ]);}
     }
 
     /**
@@ -36,6 +40,10 @@ class HolidayController extends Controller
      */
     public function create()
     {
+        if(Gate::denies('is_admin')){
+            return abort(403,'Access Denied, Only Admin Can Access');
+        }
+        elseif(Gate::allows('is_admin')) {
         $user = Auth::user();
 
         return view('masterData.holiday.add', [
@@ -43,7 +51,7 @@ class HolidayController extends Controller
             'profile_photo'=>$user->profile_photo,
             'email'=>$user->email,
             'id'=>$user->id
-        ]);
+        ]);}
     }
 
     /**
@@ -54,12 +62,15 @@ class HolidayController extends Controller
      */
     public function store(Request $request)
     {   
+        if(Gate::denies('is_admin')){
+            return abort(403,'Access Denied, Only Admin Can Access');
+        }
+        elseif(Gate::allows('is_admin')) {
         $request->validate([
             'information' => 'required',
             'start' => 'required',
             'end' => 'required',
         ]);
-        
         $start_date = date_create($request->start);
         $end_date = date_create($request->end);
         $interval = date_diff($start_date,$end_date);
@@ -81,27 +92,15 @@ class HolidayController extends Controller
         
         Alert::success('Berhasil!', 'Hari libur baru telah ditambahkan!');
         return redirect('/admin/holiday');
+        }
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit(MasterHoliday $holiday)
     {
+        if(Gate::denies('is_admin')){
+            return abort(403,'Access Denied, Only Admin Can Access');
+        }
+        elseif(Gate::allows('is_admin')) {
         $user = Auth::user();
         return view('masterdata.holiday.edit',[
             'holiday' => $holiday,
@@ -109,7 +108,7 @@ class HolidayController extends Controller
             'profile_photo'=>$user->profile_photo,
             'email'=>$user->email,
             'id'=>$user->id
-        ]);
+        ]);}
     }
 
     /**
@@ -121,6 +120,10 @@ class HolidayController extends Controller
      */
     public function update(MasterHoliday $holiday, Request $request)
     {
+        if(Gate::denies('is_admin')){
+            return abort(403,'Access Denied, Only Admin Can Access');
+        }
+        elseif(Gate::allows('is_admin')) {
         $request->validate([
             'information' => 'required',
             'date' => 'required|unique:master_holidays,date',
@@ -132,7 +135,7 @@ class HolidayController extends Controller
                 'date' => $request->date
             ]);
         Alert::success('Berhasil!', 'Hari Libur ' . $holiday->information . ' berhasil diperbaharui!');
-        return redirect('/admin/holiday');
+        return redirect('/admin/holiday');}
     }
 
     /**
@@ -143,10 +146,14 @@ class HolidayController extends Controller
      */
     public function destroy(Request $request)
     {
+        if(Gate::denies('is_admin')){
+            return abort(403,'Access Denied, Only Admin Can Access');
+        }
+        elseif(Gate::allows('is_admin')) {
         foreach ($request->selectid as $item) {
             MasterHoliday::where('id','=',$item)->delete();
         }
         Alert::success('Berhasil!', 'Hari libur yang dipilih berhasil dihapus!');
-        return redirect('/admin/holiday');
+        return redirect('/admin/holiday');}
     }
 }
